+++ tests/debugger_test.py
@@ -21,7 +21,6 @@ import unittest
 
 from absl.testing import absltest
 import jax
-from jax.experimental import pjit
 from jax._src import debugger
 from jax._src import test_util as jtu
 import jax.numpy as jnp
@@ -276,7 +275,7 @@ class CliDebuggerTest(jtu.JaxTestCase):
     jax.effects_barrier()
     self.assertRegex(stdout.getvalue(), expected)
 
-  def test_debugger_works_with_pjit(self):
+  def test_debugger_works_with_jit(self):
     if jax.default_backend() != "tpu":
       raise unittest.SkipTest("`pjit` doesn't work with CustomCall.")
 
@@ -290,18 +289,19 @@ class CliDebuggerTest(jtu.JaxTestCase):
     def g(x):
       y = f(x)
       return jnp.exp(y)
-    g = pjit.pjit(
+    g = jax.jit(
         g,
         in_shardings=jax.sharding.PartitionSpec("dev"),
         out_shardings=jax.sharding.PartitionSpec("dev"),
     )
-    with jax.sharding.Mesh(np.array(jax.devices()), ["dev"]):
-      arr = (1 + jnp.arange(8)).astype(np.int32)
+    arr = (1 + jnp.arange(8)).astype(np.int32)
+    arr2 = jnp.arange(8, dtype=jnp.int32)
+    with jax.set_mesh(jax.sharding.Mesh(np.array(jax.devices()), ["dev"])):
       expected = _format_multiline(r"""
       Entering jdb:
       \(jdb\) {}
       \(jdb\) """.format(re.escape(repr(arr))))
-      g(jnp.arange(8, dtype=jnp.int32))
+      g(arr2)
       jax.effects_barrier()
       self.assertRegex(stdout.getvalue(), expected)