+++ tests/debugging_primitives_test.py
@@ -19,7 +19,6 @@ import unittest
 from absl.testing import absltest, parameterized
 import jax
 from jax import lax
-from jax.experimental import pjit
 from jax.interpreters import pxla
 from jax._src import ad_checkpoint
 from jax._src import config
@@ -833,15 +832,15 @@ class DebugPrintParallelTest(jtu.JaxTestCase):
       jax.effects_barrier()
     self._assertLinesEqual(output(), "hello: 0\nhello: 1\nhello: 2\nhello: 3\n")
 
-  def test_unordered_print_with_pjit(self):
+  def test_unordered_print_with_jit(self):
     def f(x):
       debug_print("{}", x, ordered=False)
       return x
     mesh = jax.sharding.Mesh(np.array(jax.devices()), ['dev'])
     spec = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec('dev'))
     out_spec = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec())
-    f = pjit.pjit(f, in_shardings=spec, out_shardings=out_spec)
-    with mesh:
+    f = jax.jit(f, in_shardings=spec, out_shardings=out_spec)
+    with jax.set_mesh(mesh):
       with jtu.capture_stdout() as output:
         f(np.arange(8, dtype=jnp.int32))
         jax.effects_barrier()
@@ -851,24 +850,24 @@ class DebugPrintParallelTest(jtu.JaxTestCase):
       y = x.dot(x)
       debug_print("{}", y, ordered=False)
       return y
-    f2 = pjit.pjit(f2, in_shardings=spec, out_shardings=out_spec)
-    with jax.sharding.Mesh(np.array(jax.devices()), ['dev']):
+    f2 = jax.jit(f2, in_shardings=spec, out_shardings=out_spec)
+    with jax.set_mesh(mesh):
       with jtu.capture_stdout() as output:
         f2(np.arange(8, dtype=jnp.int32))
         jax.effects_barrier()
       self.assertEqual(output(), "140\n")
 
-  def test_nested_pjit_debug_print(self):
+  def test_nested_jit_debug_print(self):
     def f(x):
       debug_print("{}", x)
       return x
 
     with jtu.capture_stdout() as output:
-      pjit.pjit(pjit.pjit(f))(jnp.arange(8))
+      jax.jit(jax.jit(f))(jnp.arange(8))
       jax.effects_barrier()
     self.assertEqual(output(), "[0 1 2 3 4 5 6 7]\n")
 
-  def test_unordered_print_of_pjit_of_while(self):
+  def test_unordered_print_of_jit_of_while(self):
     def f(x):
       def cond(carry):
         i, *_ = carry
@@ -882,8 +881,8 @@ class DebugPrintParallelTest(jtu.JaxTestCase):
 
     mesh = jax.sharding.Mesh(np.array(jax.devices()), ['dev'])
     spec = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec('dev'))
-    f = pjit.pjit(f, in_shardings=spec, out_shardings=spec)
-    with mesh:
+    f = jax.jit(f, in_shardings=spec, out_shardings=spec)
+    with jax.set_mesh(mesh):
       with jtu.capture_stdout() as output:
         f(np.arange(8, dtype=jnp.int32))
         jax.effects_barrier()
@@ -1173,7 +1172,7 @@ class VisualizeShardingTest(jtu.JaxTestCase):
 
 class InspectShardingTest(jtu.JaxTestCase):
 
-  def test_inspect_sharding_is_called_in_pjit(self):
+  def test_inspect_sharding_is_called_in_jit_sharded(self):
 
     if jtu.is_cloud_tpu():
       raise unittest.SkipTest("Inspect sharding is not supported on libtpu.")
@@ -1192,8 +1191,8 @@ class InspectShardingTest(jtu.JaxTestCase):
     mesh = jax.sharding.Mesh(np.array(jax.devices()), ['dev'])
     spec = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec('dev'))
     out_spec = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec())
-    f = pjit.pjit(f, in_shardings=spec, out_shardings=out_spec)
-    with mesh:
+    f = jax.jit(f, in_shardings=spec, out_shardings=out_spec)
+    with jax.set_mesh(mesh):
       f(np.arange(8, dtype=jnp.int32))
     self.assertTrue(is_called)
 
@@ -1236,23 +1235,6 @@ class InspectShardingTest(jtu.JaxTestCase):
 
     f(arr)
 
-  def test_inspect_sharding_3d_pjit(self):
-    def _cb(sd):
-      self.assertIsInstance(sd, jax.sharding.NamedSharding)
-      self.assertLen(sd.device_set, 2)
-
-    def f_(x):
-      debugging.inspect_array_sharding(x, callback=_cb)
-      return jnp.square(x)
-
-    f = pjit.pjit(f_)
-    mesh = jtu.create_mesh((2,), ('x'))
-    s = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec('x'))
-    arr = jax.device_put(np.arange(8).reshape(2, 2, 2), s)
-
-    with mesh:
-      f(arr)
-
 
 def _get_output_set(output, num_lines):
   """Return a set of strings where each string is num_lines."""
@@ -1283,13 +1265,13 @@ class PartitionedDebugCallbackTest(jtu.JaxTestCase):
     def f_(x):
       debug_print("hello: {x}", x=x, partitioned=True)
 
-    f = pjit.pjit(f_)
+    f = jax.jit(f_)
     mesh = jtu.create_mesh((1, 1, 2,), ("x", "y", "z"))
     s = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec("x", "y", "z"))
     arr = jax.device_put(np.arange(24).reshape(2, 3, 4), s)
 
     with jtu.capture_stdout() as output:
-      with mesh:
+      with jax.set_mesh(mesh):
         f(arr)
       jax.effects_barrier()
 
@@ -1318,14 +1300,14 @@ class PartitionedDebugCallbackTest(jtu.JaxTestCase):
     def f_(x):
       debug_print("hello: {}", x, partitioned=True)
 
-    f = pjit.pjit(f_)
+    f = jax.jit(f_)
     mesh = jtu.create_mesh((1, 1, 2), ("x", "y", "z"))
     s = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec("x", "y", "z"))
     arr = np.arange(24).reshape(2, 3, 4)
     arr = jax.device_put(arr, s)
 
     with jtu.capture_stdout() as output:
-      with mesh:
+      with jax.set_mesh(mesh):
         f(arr)
       jax.effects_barrier()
 
@@ -1355,7 +1337,7 @@ class PartitionedDebugCallbackTest(jtu.JaxTestCase):
     def f_(x, y):
       debug_print("hello: {} {}", x, y, partitioned=True)
 
-    f = pjit.pjit(f_)
+    f = jax.jit(f_)
     mesh = jtu.create_mesh((2,), ("x"))
     s = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec("x"))
     x = np.arange(4).reshape(2, 2)
@@ -1364,7 +1346,7 @@ class PartitionedDebugCallbackTest(jtu.JaxTestCase):
     y = jax.device_put(y, s)
 
     with jtu.capture_stdout() as output:
-      with mesh:
+      with jax.set_mesh(mesh):
         f(x, y)
       jax.effects_barrier()
 
@@ -1383,14 +1365,14 @@ class PartitionedDebugCallbackTest(jtu.JaxTestCase):
     def f_(x):
       debug_print("hello: {}", x, partitioned=True)
 
-    f = pjit.pjit(f_)
+    f = jax.jit(f_)
     mesh = jtu.create_mesh((1, 1, 2), ("x", "y", "z"))
     s = jax.sharding.NamedSharding(mesh, jax.sharding.PartitionSpec("x", "y", "z"))
     arr = np.arange(24).reshape(2, 3, 4)
     arr = jax.device_put(arr, s)
 
     with jtu.capture_stdout() as output:
-      with mesh:
+      with jax.set_mesh(mesh):
         f(arr)
       jax.effects_barrier()