+++ tests/pjit_test.py
@@ -1675,7 +1675,7 @@ class ArrayPjitTest(jtu.JaxTestCase):
       f = pjit(lambda x: x,
                 in_shardings=NamedSharding(global_mesh, P('x')))
       err_msg = re.compile(
-          "Sharding passed to pjit does not match the sharding on the "
+          "Sharding passed to jit does not match the sharding on the "
           r"respective arg.*arg shape.*\[8,2\]", re.M | re.S)
       with self.assertRaisesRegex(ValueError, err_msg):
         f(input_array)
@@ -1844,7 +1844,7 @@ class ArrayPjitTest(jtu.JaxTestCase):
                 in_shardings=NamedSharding(global_mesh, P(None)))
       with self.assertRaisesRegex(
           ValueError,
-          ('Sharding passed to pjit does not match the sharding on the '
+          ('Sharding passed to jit does not match the sharding on the '
             'respective arg')):
         f(input_data, a1)